add p2p test (#76)
* init

* complete

* add comments

* add comments

* resolve comments

* change comments
