--- cmd/node/cmd/p2pstart.toml
@@ -1,8 +1,0 @@
-StaticNodes=["snode://0548d0b1a3297fea072284f86b9fd39a9f1273c46fba8951b62de5b95cd3dd846278057ec4df598a0b089a0bdc0c8fd3aa601cf01a9f30a60292ea0769388d1f@182.87.223.29:39007"]
-
-[P2PConfig]
-Name="test1"
-ECDSAKey="0x445e92837140929b190e89818c39223d1d2b9c07388d80e907adf2e3ba187563"
-KadPort="39007"
-ListenAddr="0.0.0.0:39007"
-
