+++ cmd/node/cmd/peer2.toml
@@ -1,0 +1,7 @@
+StaticNodes=["snode://607a2d7b0d1e899fb3cd3ac2ece65acd888a5de59ab0a215a8533f59c46245f60a6d70766f71738b7b2186b9302f7d1ca1430c082502ec9875ad1d3ea1ff1e29@127.0.0.1:39007"]
+
+[P2PConfig]
+Name="test2"
+ECDSAKey="0x445e92837140929b190e89818c39223d1d2b9c07388d80e907adf2e3ba187522"
+KadAddr="127.0.0.1:39008"
+ListenAddr="0.0.0.0:39008"
