+++ p2p/protocol.go
@@ -22,6 +22,7 @@ type Protocol struct {
 
 	// Version should contain the version number of the protocol.
 	Version uint
+
 	// AddPeerCh a peer joins protocol, SubProtocol should handle the channel
 	AddPeerCh chan *Peer
 
@@ -32,7 +33,7 @@ type Protocol struct {
 	ReadMsgCh chan *Message
 
 	// IsSelfConnect if value is true, means SubProtocol maintains peers itself, for example when to initiate a tcp connection.
-	// Otherwise, all SubProtols share one tcp connection which is initiated by p2p.scheduleTasks
+	// Otherwise, all SubProtols share one tcp connection which is initiated by p2p.syncPeerNodes
 	IsSelfConnect bool
 }
 
@@ -54,12 +55,4 @@ type Cap struct {
 
 func (cap Cap) String() string {
 	return fmt.Sprintf("%s/%d", cap.Name, cap.Version)
-}
-
-type capsByNameAndVersion []Cap
-
-func (cs capsByNameAndVersion) Len() int      { return len(cs) }
-func (cs capsByNameAndVersion) Swap(i, j int) { cs[i], cs[j] = cs[j], cs[i] }
-func (cs capsByNameAndVersion) Less(i, j int) bool {
-	return cs[i].Name < cs[j].Name || (cs[i].Name == cs[j].Name && cs[i].Version < cs[j].Version)
 }
